requirements.txt
PyQt6
pyserial
pyinstaller